+++ python/sglang/srt/environ.py
@@ -312,7 +312,7 @@ class Envs:
     SGLANG_VLM_CACHE_SIZE_MB = EnvInt(100)
     SGLANG_IMAGE_MAX_PIXELS = EnvInt(16384 * 28 * 28)
     SGLANG_RESIZE_RESAMPLE = EnvStr("")
-    SGLANG_ENABLE_MM_BUFFER = EnvBool(True)
+    SGLANG_MM_BUFFER_SIZE = EnvInt(0)
 
     # Release & Resume Memory
     SGLANG_MEMORY_SAVER_CUDA_GRAPH = EnvBool(False)
